+++ src/acbm/assigning/primary_feasible.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 import pandas as pd
 from pandarallel import pandarallel
 
@@ -283,118 +281,3 @@ def _get_possible_zones(
         .apply(list)
         .to_dict()
     )
-
-
-def fill_missing_zones(
-    activity: pd.Series,
-    travel_times_est: dict,
-    activities_per_zone: pd.DataFrame,
-    activity_col: str,
-    use_mode: bool = False,
-):
-    """
-    This function fills in missing zones in the activity chains. It uses a travel time matrix based on euclidian distance instead of
-    the computed travel time matrix which is a sparse matrix.
-
-    Parameters
-    ----------
-    activity: pd.Series
-        A row from the activity chains dataframe
-    travel_times_est: dict
-        A dictionary with keys as tuples ({id_col}_from, {id_col}_to) and values as dictionaries containing time estimates.
-        It is the output of zones_to_time_matrix()
-    activities_per_zone: pd.DataFrame
-        A dataframe with the number of activities and floorspace for each zone. The columns are 'OA21CD', 'counts', 'floor_area', 'activity'
-        where 'activity' is the activity type as defined in the osmox config file
-    activity_col: str
-        The column name for the activity type
-    use_mode: bool
-        If True, the function will use the mode of transportation to estimate the travel time: time_{mode}.
-        If False, it will use the average travel time: time_average.
-        Default is False.
-
-    Returns
-    -------
-    str
-        The zone that has the estimated time closest to the given time. The zone also has an activity that matches the activity_col value.
-
-    """
-    activity_purpose = activity[activity_col]
-    from_zone = activity["OA21CD"]
-    to_zones = activities_per_zone[activities_per_zone["activity"] == activity_purpose][
-        "OA21CD"
-    ].tolist()
-
-    logger.debug(
-        f"Activity {activity.TripID} | person: {activity.id}: Number of possible destination zones: {len(to_zones)}"
-    )
-
-    time = activity["TripTotalTime"]
-
-    mode = activity["mode"] if use_mode else None
-
-    return _get_zones_using_time_estimate(
-        estimated_times=travel_times_est,
-        from_zone=from_zone,
-        to_zones=to_zones,
-        time=time,
-        mode=mode,
-    )
-
-
-def _get_zones_using_time_estimate(
-    estimated_times: dict,
-    from_zone: str,
-    to_zones: list,
-    time: int,
-    mode: Optional[str] = None,
-) -> str:
-    """
-    This function returns the zone that has the estimated time closest to the given time. It is meant to be used inside fill_missing_zones()
-
-    Parameters:
-    ----------
-    estimated_times: dict
-        A dictionary with keys as tuples ({id_col}_from, {id_col}_to) and values as dictionaries containing time estimates. It is the output of zones_to_time_matrix()
-    id_col: str
-        The column name for the zone id.
-    from_zone: str
-        The zone of the previous activity
-    to_zones (list): A list of destination zones.
-    time:
-        The target time to compare the estimated times with.
-    mode: str, optional
-        The mode of transportation. It should be one of ['car', 'pt', 'walk', 'cycle']. If not provided, the function uses 'time_average'.
-
-    Returns
-    -------
-    str
-        The zone that has the estimated time closest to the given time.
-    """
-
-    acceptable_modes = ["car", "pt", "walk", "cycle"]
-
-    if mode is not None and mode not in acceptable_modes:
-        error_message = f"Invalid mode: {mode}. Mode must be one of {acceptable_modes}."
-        raise ValueError(error_message)
-
-    # Convert to_zones to a set for faster lookup
-    to_zones_set = set(to_zones)
-
-    # Filter the entries where {id_col}_from matches the specific zone and {id_col}_to is in the specific list of zones
-    filtered_dict = {
-        k: v
-        for k, v in estimated_times.items()
-        if k[0] == from_zone and k[1] in to_zones_set
-    }
-    # get to_zone where time_average is closest to "time"
-    if mode is not None:
-        closest_to_zone = min(
-            filtered_dict.items(), key=lambda item: abs(item[1][f"time_{mode}"] - time)
-        )
-    else:
-        closest_to_zone = min(
-            filtered_dict.items(), key=lambda item: abs(item[1]["time_average"] - time)
-        )
-
-    return closest_to_zone[0][1]
